@@ -11,7 +11,7 @@
     "logs": "firebase functions:log"
   },
   "engines": {
-    "node": "16"
+    "node": "22"
   },
   "main": "lib/index.js",
   "dependencies": {
@@ -20,7 +20,7 @@
     "cors": "^2.8.5",
     "express": "^4.18.2",
     "firebase-admin": "^11.5.0",
-    "firebase-functions": "^4.2.0",
+    "firebase-functions": "^5.1.0",
     "moment": "^2.29.4"
   },
   "devDependencies": {
