@@ -1,21 +1,30 @@
-import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs';
+import { Injectable } from "@angular/core";
+import { Observable } from "rxjs";
 import {
-  addDoc, collection, collectionData, CollectionReference, deleteDoc,
-  doc, docData, DocumentData, Firestore, Timestamp, updateDoc
-} from '@angular/fire/firestore'
-import { HttpClient } from '@angular/common/http';
+  addDoc,
+  collection,
+  collectionData,
+  CollectionReference,
+  deleteDoc,
+  doc,
+  docData,
+  DocumentData,
+  Firestore,
+  Timestamp,
+  updateDoc,
+} from "@angular/fire/firestore";
+import { HttpClient } from "@angular/common/http";
 
 import { environment } from "src/environments/environment";
-import { map } from 'rxjs/operators';
+import { map } from "rxjs/operators";
 interface Tickets {
   CATEGORY: string;
   STATUS: string;
-  DUE_DATE:string;
-  CREATE_DATE:string;
-  LOG:string;
+  DUE_DATE: string;
+  CREATE_DATE: string;
+  LOG: string;
   HEADLINE: string;
-  CURRENT_HANDLER_USER_ID:string;
+  CURRENT_HANDLER_USER_ID: string;
   TICKET_ID: string;
   DETAIL: string;
   OPENER_USER_ID: string;
@@ -23,13 +32,13 @@ interface Tickets {
 }
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: "root",
 })
 export class TicketService {
   private ticketCollection: CollectionReference<DocumentData>;
 
   constructor(private firestore: Firestore, private _http: HttpClient) {
-    this.ticketCollection = collection(firestore, 'tickets');
+    this.ticketCollection = collection(firestore, "tickets");
   }
 
   create(tickets: Tickets) {
@@ -37,28 +46,31 @@ export class TicketService {
   }
   getAll() {
     return collectionData(this.ticketCollection, {
-      idField: 'id',
+      idField: "id",
     }) as Observable<Tickets[]>;
   }
 
   get(id: string) {
     const ticketDocumentReference = doc(this.firestore, `tickets/${id}`);
-    return docData(ticketDocumentReference, { idField: 'id' });
+    return docData(ticketDocumentReference, { idField: "id" });
   }
 
   update(tickets: Tickets) {
-
-    const ticketDocumentReference = doc(this.firestore, `tickets/${tickets.id}`);
+    const ticketDocumentReference = doc(
+      this.firestore,
+      `tickets/${tickets.id}`
+    );
     return updateDoc(ticketDocumentReference, { ...tickets });
   }
-
 
   delete(id: string) {
     const ticketDocumentReference = doc(this.firestore, `tickets/${id}`);
     return deleteDoc(ticketDocumentReference);
   }
   async getAllTicketFromAPI() {
-    const response = await fetch("https://us-central1-fvl-app.cloudfunctions.net/api/service");
+    const response = await fetch(
+      "https://us-central1-foodvilla-1fe60.cloudfunctions.net/api/service"
+    );
     const users = await response.json();
     return users.data;
   }
