@@ -12,7 +12,6 @@ import {
   Firestore,
   setDoc,
   updateDoc,
-
 } from "@angular/fire/firestore";
 
 interface Customers {
@@ -25,7 +24,7 @@ interface Customers {
   minimumMoney: string;
   currentMoney: number;
 }
-interface BoothCode{
+interface BoothCode {
   code: string;
 }
 
@@ -86,9 +85,10 @@ export class CustomerService {
   }
 
   async getAllCustomerFromAPI() {
-    const response = await fetch("https://us-central1-fvl-app.cloudfunctions.net/api/customers");
+    const response = await fetch(
+      "https://us-central1-foodvilla-1fe60.cloudfunctions.net/api/customers"
+    );
     const users = await response.json();
     return users.data;
   }
-  
 }
